Remove duplicated default tiling shortcuts
XFCE for some reason is now inserting a set of default window-tiling shortcuts
under "/xfwm4/custom". They not only seem to not work but, even worse, cause
our custom shortcuts to break because there can't be two different shortcuts
mapping to the same command, or neither will work.

This commit solves that by adding a new step to the keyboard playbook to
remove all of system default shortcuts that conflict with the ones we
add by default.